# Backend/doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Newsletters",
    "version": "1.0.0",
    "description": "Documentation de l'API pour gérer les blogs, utilisateurs, abonnés, commentaires et newsletters."
  },
  "paths": {
...
,"/blogs": {
  "get": {
    "summary": "Récupérer tous les blogs",
    "tags": ["Blogs"],
    "responses": {
      "200": {
        "description": "La liste des blogs",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Blog"
              }
            }
          }
        }
      }
    }
  },
  "post": {
    "summary": "Créer un nouveau blog",
    "tags": ["Blogs"],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/Blog"
          }
        }
      }
    },
    "responses": {
      "201": {
        "description": "Blog créé avec succès",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Blog"
            }
          }
        }
      },
      "400": {
        "description": "Requête invalide"
      }
    }
  }
},
"/blogs/{id}": {
  "get": {
    "summary": "Récupérer un blog par ID",
    "tags": ["Blogs"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Blog trouvé",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Blog"
            }
          }
        }
      },
      "404": {
        "description": "Blog non trouvé"
      }
    }
  },
  "put": {
    "summary": "Mettre à jour un blog par ID",
    "tags": ["Blogs"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/Blog"
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Blog mis à jour avec succès",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Blog"
            }
          }
        }
      },
      "400": {
        "description": "Requête invalide"
      },
      "404": {
        "description": "Blog non trouvé"
      }
    }
  },
  "delete": {
    "summary": "Supprimer un blog par ID",
    "tags": ["Blogs"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "204": {
        "description": "Blog supprimé avec succès"
      },
      "404": {
        "description": "Blog non trouvé"
      }
    }
  }
},
...
,"/comments": {
  "get": {
    "summary": "Récupérer tous les commentaires",
    "tags": ["Comments"],
    "responses": {
      "200": {
        "description": "La liste des commentaires",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Comment"
              }
            }
          }
        }
      }
    }
  },
  "post": {
    "summary": "Créer un nouveau commentaire",
    "tags": ["Comments"],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/Comment"
          }
        }
      }
    },
    "responses": {
      "201": {
        "description": "Commentaire créé avec succès",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Comment"
            }
          }
        }
      },
      "400": {
        "description": "Requête invalide"
      }
    }
  }
},
"/comments/{id}": {
  "get": {
    "summary": "Récupérer un commentaire par ID",
    "tags": ["Comments"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Commentaire trouvé",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Comment"
            }
          }
        }
      },
      "404": {
        "description": "Commentaire non trouvé"
      }
    }
  },
  "put": {
    "summary": "Mettre à jour un commentaire par ID",
    "tags": ["Comments"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/Comment"
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Commentaire mis à jour avec succès",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Comment"
            }
          }
        }
      },
      "400": {
        "description": "Requête invalide"
      },
      "404": {
        "description": "Commentaire non trouvé"
      }
    }
  },
  "delete": {
    "summary": "Supprimer un commentaire par ID",
    "tags": ["Comments"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "204": {
        "description": "Commentaire supprimé avec succès"
      },
      "404": {
        "description": "Commentaire non trouvé"
      }
    }
  }
}
...
  },
  "components": {
    "schemas": {
      "Blog": {
        "type": "object",
        "required": ["userId", "content", "title"],
        "properties": {
          "userId": {
            "type": "string",
            "description": "L'ID de l'utilisateur qui a créé le blog"
          },
          "content": {
            "type": "string",
            "description": "Contenu du blog"
          },
          "title": {
            "type": "string",
            "description": "Titre du blog"
          },
          "image": {
            "type": "string",
            "description": "URL de l'image associée au blog"
          },
          "category": {
            "type": "string",
            "description": "Catégorie du blog"
          },
          "slug": {
            "type": "string",
            "description": "Slug SEO du blog"
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "required": ["userId", "content", "createdAt"],
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID de l'utilisateur qui a posté le commentaire"
          },
          "content": {
            "type": "string",
            "description": "Contenu du commentaire"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date et heure de création du commentaire"
          },
          "postId": {
            "type": "string",
            "description": "ID du poste auquel le commentaire est associé"
          }
        }
      },
...
    }
  }
}